+++ src/handlers/index.ts
@@ -1,0 +1,13 @@
+import getAll from './getAll';
+import createUser from './create';
+import getById from './getById';
+import remove from './remove';
+import update from './update';
+
+export default {
+  getAll,
+  createUser,
+  getById,
+  remove,
+  update,
+};
